requirements.txt
elementpath==1.4.0
numpy==1.18.1
opencv-python==4.1.2.30
pandas==0.25.3
pdf2image==1.11.0
Pillow==7.0.0
requests==2.20.0
xlrd==1.2.0
